dependencies: update to fix vulnerability issues
The dependencies updated were:

- eslint: 4.5.0 -> 5.13.0
- eslint-plugin-import: 2.7.0 -> 2.16.0
- mocha: 3.5.0 -> 5.2.0

Even though most of the updates where major, nothing
broke. The only problem is that when the linter
runs, a deprecation warning is shown that we can't
fix until `airbnb` fixes it.

This is the output:

(node:11013) [ESLINT_LEGACY_OBJECT_REST_SPREAD] DeprecationWarning: The 'parserOptions.ecmaFeatures.experimentalObjectRestSpread' option is deprecated. Use 'parserOptions.ecmaVersion' instead. (found in "airbnb-base")